+++ client/CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Input selection can select basic outputs without ed25519 address in the address unlock condition;
 - Automatic input selection with sender and issuer features;
 - Automatic alias governance transition does not change output amount in input selection;
+- ISA does not select aliases it can't unlock for amount;
 
 ## 2.0.1-rc.6 - 2023-02-08